requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pymongo==4.6.1
pydantic==2.6.1
google-generativeai==0.3.2
python-dotenv==1.0.1
pytest==8.0.0
black==24.1.1
isort==5.13.2
mypy==1.8.0
python-multipart==0.0.9
httpx==0.26.0
tenacity==8.2.3
structlog==24.1.0

# Testing dependencies
pytest-cov==4.1.0
pytest-asyncio==0.23.5
pytest-mock==3.12.0
pytest-xdist==3.5.0
coverage==7.4.1
pydantic-settings==2.1.0 